Term Project Code/requirements.txt
openai>=1.3.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
faiss-cpu==1.7.4
numpy==1.24.3
tqdm==4.66.1
PyMuPDF==1.23.0
arxiv==1.4.7
pandas==2.0.3
aiohttp==3.8.5
aiofiles==23.2.1
requests==2.31.0
rouge-score==0.1.2